Don't abuse 'stdint.h' types

+++ src/generate_entropy.c
@@ -16,6 +16,7 @@
  * this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
+#include <stddef.h>
 #include <stdint.h>
 #include <stdlib.h>
 #include <math.h> /* log2() */
@@ -49,7 +50,7 @@ static inline Image* alloc_and_init_image(const Args* args, ByteArray* bytes) {
  * https://8dcc.github.io/programming/understanding-entropy.html
  */
 static double entropy(void* data, size_t data_sz) {
-    uint32_t* occurrences = calloc(256, sizeof(uint32_t));
+    size_t* occurrences = calloc(256, sizeof(size_t));
 
     /* Count the occurrences of each byte in the input */
     for (size_t i = 0; i < data_sz; i++) {
